TensorBranching

Documentation for TensorBranching.

Quick start

Install

Press ] to enter the package mode in Julia Repl, and type the following command to install the package:

pkg> add https://github.com/ArrogantGao/TensorBranching.jl

solve the maximum independent set problem

Currently we provide a simple interface to solve the maximum independent set problem. You can use the following code:

julia> g = random_regular_graph(60, 3)
{60, 90} undirected simple Int64 graph

julia> missolve(g, strategy = SetCoverBranching(), kneighbor = 1, show_count = true, usr_rv = true)
(26, 206)

where kneighbor for layer of neighbor considered, the first result is the size of the maximum independent set, and the second result is the number of branches visited.

The standard mis1 and mis2 strategies are also available.

julia> using TensorBranching: counting_mis1, counting_mis2

julia> using EliminateGraphs

julia> counting_mis1(EliminateGraph(g))
CountingMIS(26, 58148786)

julia> counting_mis2(EliminateGraph(g))
CountingMIS(26, 188)

Branching Tree

To check the branching process in detail, we store all the branching process as tree.

julia> g = random_regular_graph(20, 3)
{20, 30} undirected simple Int64 graph

julia> tree, mis = branching_tree(g, SetCoverBranching(), 1, use_rv = true)
(G{20}
├─ G{16}
│  └─ G{14}
│     ├─ G{11}
│     │  └─ G{9}
│     │     └─ G{7}
│     │        ⋮
│     │        
│     └─ G{7}
│        └─ G{5}
│           └─ G{3}
│              ⋮
│              
├─ G{12}
│  └─ G{10}
│     └─ G{8}
│        ├─ G{5}
│        │  └─ G{3}
│        │     ⋮
│        │     
│        └─ G{3}
│           └─ G{0}
└─ G{16}
   └─ G{14}
      ├─ G{11}
      │  └─ G{9}
      │     └─ G{7}
      │        ⋮
      │        
      └─ G{8}
         └─ G{6}
            └─ G{4}
               ⋮
               
, 6)

and the graph of current state, the vertices to be removed are stored:

julia> tree.graph
{20, 30} undirected simple Int64 graph

julia> tree.removed
3-element Vector{Vector{Int64}}:
 [1, 5, 13, 20]
 [1, 5, 13, 10, 15, 20, 4, 16]
 [1, 5, 12, 16]

the vertices to be removed correspond to the children of the current node.

Detailed documentation

TensorBranching.BranchingNode

TensorBranching.BranchingTable

TensorBranching.Clause

TensorBranching.NaiveBranching

TensorBranching.SetCoverBranching

TensorBranching.SubCover

TensorBranching.branching_tree

TensorBranching.clause

TensorBranching.complexity

TensorBranching.counting_mis1

TensorBranching.counting_mis2

TensorBranching.cover

TensorBranching.covered_by

TensorBranching.covered_by

TensorBranching.covered_items

TensorBranching.env_filter

TensorBranching.missolve

TensorBranching.optimal_branching_dnf

TensorBranching.setcover_strategy

TensorBranching.subcovers

TensorBranching.subcovers_naive

TensorBranching.BranchingNode — Type

mutable struct BranchingNode

A mutable struct representing a node in a branching tree.

Fields

children::Vector{BranchingNode}: The children nodes of the current node.

graph::SimpleGraph{Int}: The graph associated with the current node.

removed::Vector{Vector{Int}}: The removed edges from the graph.

Constructor

BranchingNode(graph::SimpleGraph{Int}; children::Vector{BranchingNode} = Vector{BranchingNode}(), removed::Vector{Vector{Int}} = Vector{Vector{Int}}()): Constructs a new BranchingNode object.

TensorBranching.BranchingTable — Type

BranchingTable{N, C}

A table of branching configurations. The table is a vector of vectors of StaticBitVector{N, C}. Type parameters are:

N: The number of bits in the bit vector.

C: The number of integers as the storage.

Examples

julia> graph_sat = graph_from_tuples(3, [(1, 2), (2, 3), (1, 3)])
{3, 3} undirected simple Int64 graph

julia> tbl = reduced_alpha_configs(graph_sat, [1, 2])
BranchingTable{N}
001

To cover the branching table, at least one clause in each row must be satisfied.

TensorBranching.Clause — Type

Clause{N, T}

A clause is a pair of bit strings, mask and val, where mask is a bit string that indicates the bits that are relevant to the clause, and val is a bit string that indicates the bits that must be satisfied. The clause is satisfied if and only if val is covered by the bit string and mask.

N: The number of bits in the bit vector.

T: The number of integers as the storage.

Examples

julia> Clause(bit"1110", bit"1001")
Clause{4, 3, Int64}: mask: 1110 ₍₂₎, val: 1000 ₍₂₎

If some bit in mask is set as 0, then the corresponding bit in val must be 0.

TensorBranching.NaiveBranching — Type

struct NaiveBranching <: BranchingStrategy

A struct representing the NaiveBranching branching strategy.

TensorBranching.SetCoverBranching — Type

struct SetCoverBranching <: BranchingStrategy

A struct representing a branching strategy for set cover problems.

Fields

max_itr::Int: The maximum number of iterations.

Constructors

SetCoverBranching(): Constructs a SetCoverBranching object with a default value of 2 for max_itr.

SetCoverBranching(max_itr::Int): Constructs a SetCoverBranching object with the specified max_itr value.

TensorBranching.SubCover — Type

SubCover{N, T}

A subcover is a pair of a set of integers ids, a clause clause and a integer n_rm. The ids for the truth covered by the clause, and n_rm is the number of vertices to remove.

N: The number of bits in the bit vector.

T: The number of integers as the storage.

Examples

julia> SubCover([1, 2], Clause(bit"1110", bit"1001"), 3)
SubCover{4, Int64}: ids: Set([2, 1]), mask: 1110 ₍₂₎, val: 1000 ₍₂₎, n_rm: 3

TensorBranching.branching_tree — Method

branching_tree(g::SimpleGraph, strategy::BranchingStrategy, kneighbor::Int, use_rv::Bool)

Constructs a branching tree based on the given graph g using the specified strategy. The kneighbor parameter determines the number of neighbors to consider during branching. If use_rv is true, number of vertex removed is used; otherwise, only count the clause.

Arguments

g::SimpleGraph: The input graph.

strategy::BranchingStrategy: The branching strategy to use.

kneighbor::Int: The number of neighbors to consider during branching.

use_rv::Bool: Whether to use number of vertex removed.

Returns

tree: The constructed branching tree.

branch_num: The number of branches in the tree.

TensorBranching.clause — Method

clause(bitstrings::AbstractVector{BitStr{N, T}})

Return a clause that covers all the bit strings.

TensorBranching.complexity — Method

complexity(sbranches::Vector{Int})

Compute the complexity of a set of branches.

This function solves the equation 1 = sum(x^(-i) for i in sbranches) for x, where sbranches is a vector of integers representing the branches.

Arguments

sbranches::Vector{Int}: A vector of integers representing the branches.

Returns

Float64: The value of x that satisfies the equation.

TensorBranching.counting_mis1 — Method

counting_mis1(eg::EliminateGraph)

Compute the size of the maximum independent set (MIS) in the given EliminateGraph and the number of branches by mis1.

Arguments

eg::EliminateGraph: The graph to compute the MIS for.

Returns

CountingMIS: The size of the maximum independent set.

TensorBranching.counting_mis2 — Method

counting_mis2(eg::EliminateGraph)

Compute the size of the maximum independent set (MIS) in the given EliminateGraph and the number of branches by mis2.

Arguments

eg::EliminateGraph: The graph to compute the MIS for.

Returns

CountingMIS: The size of the maximum independent set.

TensorBranching.cover — Method

cover(sub_covers::AbstractVector{SubCover{N, T}}; max_itr::Int = 2, min_complexity::TF = 1.0) where{N, T, TF}

The cover function performs a set covering algorithm on a collection of sub-covers.

Arguments

sub_covers::AbstractVector{SubCover{N, T}}: A collection of sub-covers.

max_itr::Int = 2: The maximum number of iterations to perform.

min_complexity::TF = 1.0: The minimum complexity threshold.

Returns

picked: The selected sub-covers.

cx: The complexity of the selected sub-covers.

TensorBranching.covered_by — Method

covered_by(a::BitStr, b::BitStr, mask::BitStr)

Check if a is covered by b with mask. The function returns true if and only if a and b are the same when masked by mask.

TensorBranching.covered_by — Method

covered_by(a::BitStr, clause::Clause)

Check if a is covered by the clause. The function returns true if and only if a and clause.val are the same when masked by clause.mask.

TensorBranching.covered_items — Method

covered_items(bitstrings, clause::Clause)

Return the indices of the bit strings that are covered by the clause.

TensorBranching.env_filter — Method

env_filter(tbl::BranchingTable{N, C}, g::SimpleGraph, vertices::Vector{Int}, openvertices::Vector{Int}) -> BranchingTable

Filter the given tbl based on the environment of the vertices.

Arguments

tbl::BranchingTable{N, C}: The branching table to filter.

g::SimpleGraph: The graph representing the environment.

vertices::Vector{Int}: The vertices to consider.

openvertices::Vector{Int}: The open vertices.

Returns

A new BranchingTable object containing the filtered rows.

TensorBranching.missolve — Method

missolve(g::SimpleGraph; show_count = false, strategy::BranchingStrategy = NaiveBranching(), kneighbor::Int = 2, use_rv::Bool = true)

Solves the maximum independent set (MIS) problem for a given graph g using a specified branching strategy.

Arguments

g::SimpleGraph: The input graph.

show_count::Bool = false: Whether to return the count of iterations performed during the solving process.

strategy::BranchingStrategy = NaiveBranching(): The branching strategy to use. Defaults to NaiveBranching().

kneighbor::Int = 2: The number of neighbors to consider during the branching process. Defaults to 2.

use_rv::Bool = true: Whether to use number of removed vertices during the branching process. Defaults to true.

Returns

If show_count is false, returns the maximum independent set (MIS) of the graph g.

If show_count is true, returns a tuple (mis, count) where mis is the maximum independent set (MIS) of the graph g and count is the number of iterations performed during the solving process.

TensorBranching.optimal_branching_dnf — Method

optimal_branching_dnf(g::SimpleGraph, strategy::BranchingStrategy, kneighbor::Int, use_rv::Bool)

Compute the optimal branching in a directed acyclic graph (DAG) using a specified branching strategy.

Arguments

g::SimpleGraph: The input graph.

strategy::BranchingStrategy: The branching strategy to use.

kneighbor::Int: The number of neighbors to consider when selecting the branching vertex.

use_rv::Bool: Whether to use the number of removed vertices during the branching process.

Returns

vertices: The set of vertices selected for the optimal branching.

openvertices: The set of open vertices in the optimal branching.

impl_strategy: The implementation strategy used for the optimal branching.

TensorBranching.setcover_strategy — Method

setcover_strategy(tbl::BranchingTable{N, C}, vertices::Vector{Int}, g::SimpleGraph, use_rv::Bool; max_itr::Int = 1) where{N, C}

This function implements a set covering strategy for a given set of vertices in a graph.

Arguments

tbl::BranchingTable{N, C}: The branching table.

vertices::Vector{Int}: The set of vertices to cover.

g::SimpleGraph: The graph.

use_rv::Bool: A flag indicating whether to use the number of vertices removed.

max_itr::Int = 2: The maximum number of iterations.

Returns

DNF: The branching strategy.

TensorBranching.subcovers — Method

subcovers(bss::AbstractVector{Vector{BitStr{N, T}}}, vertices::Vector{Int}, g::SimpleGraph, use_rv::Bool) where {N, T}

Compute the subcovers of a set of bit strings by iteratively gathering clauses.

Arguments

bss::AbstractVector{Vector{BitStr{N, T}}}: A vector of vectors of bit strings.

vertices::Vector{Int}: A vector of integers representing vertices.

g::SimpleGraph: A simple graph.

use_rv::Bool: A boolean indicating whether to use the number of vertices removed.

Returns

allcovers::Vector{SubCover}: A vector of SubCover objects representing the subcovers.

TensorBranching.subcovers_naive — Method

subcovers_naive(bs, vertices, g, use_rv)

Compute the subcovers for a given set of bit strings by search all possible clauses.

Arguments

bs::Union{Vector{BitStr{N, T}}, AbstractVector{Vector{BitStr{N, T}}}}: The set of bit strings.

vertices::Vector{Int}: The vertices.

g::SimpleGraph: The graph.

use_rv::Bool: A flag indicating whether to use the number of vertices removed.

Returns

allcovers::Vector{SubCover{N, T}}: The computed subcovers.