API

Index

TensorBranching.AbstractBranching

TensorBranching.AbstractMeasure

TensorBranching.AbstractSetCoverSolver

TensorBranching.AbstractTruthFilter

TensorBranching.AbstractVertexSelector

TensorBranching.Branch

TensorBranching.Branches

TensorBranching.BranchingNode

TensorBranching.BranchingTable

TensorBranching.Clause

TensorBranching.CountingMIS

TensorBranching.D3Measure

TensorBranching.EnvFilter

TensorBranching.MinBoundarySelector

TensorBranching.NaiveBranching

TensorBranching.SetCoverBranching

TensorBranching.SolverConfig

TensorBranching.SubCover

TensorBranching.branching_tree

TensorBranching.clause

TensorBranching.closed_neighbors

TensorBranching.complexity

TensorBranching.count_mis

TensorBranching.counting_mis1

TensorBranching.counting_mis2

TensorBranching.cover

TensorBranching.covered_by

TensorBranching.covered_by

TensorBranching.covered_items

TensorBranching.effective_γ

TensorBranching.graph_from_artifact

TensorBranching.mis_solver

TensorBranching.neighbor_cover

TensorBranching.neighbors_2nd

TensorBranching.open_neighbors

TensorBranching.open_vertices

TensorBranching.optimal_branches

TensorBranching.removed_vertices

TensorBranching.setcover_strategy

TensorBranching.solve_mis

TensorBranching.subcovers

TensorBranching.subcovers_naive

Modules

TensorBranching.AbstractBranching — Type

abstract type AbstractBranching

Abstract type representing a branching strategy. For each branching strategy, a corresponding impl_strategy function should be implemented.

TensorBranching.AbstractMeasure — Type

abstract type AbstractMeasure

Abstract type representing a measurement of the size of given graphs. For each measurement, a corresponding measure function should be implemented.

TensorBranching.AbstractSetCoverSolver — Type

abstract type AbstractSetCoverSolver

Abstract type representing a solver for the set cover problem, including linear programming and integer programming solvers.

TensorBranching.AbstractTruthFilter — Type

abstract type AbstractTruthFilter

Abstract type representing a truth filter.

This type serves as a base for defining different truth filter strategies.

TensorBranching.AbstractVertexSelector — Type

abstract type AbstractVertexSelector

Abstract type representing a vertex selector, which selects a subset of vertices from a given graph for branching. A corresponding select_vertex function should be implemented.

TensorBranching.Branch — Type

struct Branch

A struct representing a branching strategy.

Fields

vertices_removed::Vector{Int}: A vector of integers representing the vertices removed in the branching strategy.

mis::Int: An integer representing the maximum independent set (MIS) size of the branching strategy.

TensorBranching.Branches — Type

struct Branches

A struct representing a collection of branches.

Fields

branches::Vector{Branch}: A vector of Branch objects.

TensorBranching.BranchingNode — Type

mutable struct BranchingNode

A mutable struct representing a node in a branching tree.

Fields

children::Vector{BranchingNode}: The children nodes of the current node.

graph::SimpleGraph{Int}: The graph associated with the current node.

removed::Vector{Vector{Int}}: The removed edges from the graph.

Constructor

BranchingNode(graph::SimpleGraph{Int}; children::Vector{BranchingNode} = Vector{BranchingNode}(), removed::Vector{Vector{Int}} = Vector{Vector{Int}}()): Constructs a new BranchingNode object.

TensorBranching.BranchingTable — Type

BranchingTable{INT}

A table of branching configurations. The table is a vector of vectors of INT. Type parameters are:

INT: The number of integers as the storage.

Fields

bit_length::Int: The length of the bit string.

table::Vector{Vector{INT}}: The table of bitstrings used for branching.

To cover the branching table, at least one clause in each row must be satisfied.

TensorBranching.Clause — Type

Clause{INT <: Integer}

A clause is a pair of bit strings, mask and val, where mask is a bit string that indicates the bits that are relevant to the clause, and val is a bit string that indicates the bits that must be satisfied. The clause is satisfied if and only if val is covered by the bit string and mask.

INT: The number of integers as the storage.

Examples

julia> Clause(bit"1110", bit"1001")
Clause{DitStr{2, 4, Int64}}: mask: 1110 ₍₂₎, val: 1000 ₍₂₎

If some bit in mask is set as 0, then the corresponding bit in val must be 0.

TensorBranching.CountingMIS — Type

struct CountingMIS

A struct representing a counting maximum independent set (MIS), where mis for size of the maximum independent set and count for the count of branches.

Fields

mis::Int: The maximum independent set.

count::Int: The count of the maximum independent set.

Constructors

CountingMIS(mis::Int): Constructs a CountingMIS object with the given maximum independent set and count set to 1.

CountingMIS(mis::Int, count::Int): Constructs a CountingMIS object with the given maximum independent set and count.

Examples

julia> CountingMIS(2, 2) + 2
CountingMIS(4, 2)

julia> CountingMIS(1, 1) + CountingMIS(2, 2)
CountingMIS(3, 3)

julia> max(CountingMIS(1, 1), CountingMIS(2, 2))
CountingMIS(2, 3)

TensorBranching.D3Measure — Type

D3Measure <: AbstractMeasure

A measure of complexity by counting the number of vertices with degree at least 3.

TensorBranching.EnvFilter — Type

struct EnvFilter <: AbstractTruthFilter

A struct representing an environment filter, using the information of the environment to filter out some configurations which can not reach the optimal mis solution.

TensorBranching.MinBoundarySelector — Type

struct MinBoundarySelector <: AbstractVertexSelector

The MinBoundarySelector struct represents a strategy for selecting a subgraph with the minimum number of open vertices by k-layers of neighbors.

Fields

k::Int: The number of layers of neighbors to consider when selecting the subgraph.

TensorBranching.NaiveBranching — Type

struct NaiveBranching <: AbstractBranching

A struct representing the NaiveBranching branching strategy.

TensorBranching.SetCoverBranching — Type

struct SetCoverBranching <: AbstractBranching

A struct representing a branching strategy for set cover problems.

Fields

max_itr::Int: The maximum number of iterations.

Constructors

SetCoverBranching(): Constructs a SetCoverBranching object with a default value of 2 for max_itr.

SetCoverBranching(max_itr::Int): Constructs a SetCoverBranching object with the specified max_itr value.

verbose::Bool: A boolean indicating whether to print the output.

TensorBranching.SolverConfig — Type

SolverConfig{ST<:AbstractMISSolver, BT<:AbstractBranching, MT<:AbstractMeasure, VT<:AbstractVertexSelector, FT<:AbstractTruthFilter}

The configuration for the optimal branching MIS solver.

Fields

table_solver::ST: the MIS solver, e.g., TensorNetworkSolver()

branching_strategy::BT: the branching strategy, e.g., NaiveBranching(), SetCoverBranching()

measure::MT: the measure for the branching strategy, e.g., NumOfVertices(), D3Measure()

vertex_selector::VT: the vertex selector, e.g., MinBoundarySelector(2), ManualSelector([1, 2, 3])

table_filter::FT: the truth filter, e.g., EnvFilter(), NoFilter()

TensorBranching.SubCover — Type

SubCover{INT <: Integer}

A subcover is a pair of a set of integers ids, a clause clause and a integer n_rm. The ids for the truth covered by the clause, and n_rm is the number of vertices to remove.

INT: The number of integers as the storage.

Examples

julia> SubCover([1, 2], Clause(bit"1110", bit"1001"), 3)
SubCover{DitStr{2, 4, Int64}}: ids: Set([2, 1]), mask: 1110 ₍₂₎, val: 1000 ₍₂₎, n_rm: 3

TensorBranching.branching_tree — Method

branching_tree(g::SimpleGraph, config::SolverConfig)

Constructs a branching tree based on the given graph g using the specified config.

TensorBranching.clause — Method

clause(n::Int, bitstrings::AbstractVector{INT})

Return a clause that covers all the bit strings.

TensorBranching.closed_neighbors — Method

closed_neighbors(g::SimpleGraph, vertices::Vector{Int})

Returns a set of vertices that includes the input vertices as well as their open neighbors.

Arguments

g::SimpleGraph: The input graph.

vertices::Vector{Int}: The vertices for which closed neighbors are to be computed.

Returns

A set of vertices that includes the input vertices as well as their open neighbors.

TensorBranching.complexity — Method

complexity(sbranches::Vector{Int})

Compute the complexity of a set of branches.

This function solves the equation 1 = sum(x^(-i) for i in sbranches) for x, where sbranches is a vector of integers representing the branches.

Arguments

sbranches::Vector{Int}: A vector of integers representing the branches.

Returns

Float64: The value of x that satisfies the equation.

TensorBranching.count_mis — Method

count_mis(g::SimpleGraph, config::SolverConfig)

Count the number of MIS for the given graph and configuration.

Arguments

g::SimpleGraph: the input graph

config::SolverConfig: the solver configuration

TensorBranching.counting_mis1 — Method

counting_mis1(eg::EliminateGraph)

Compute the size of the maximum independent set (MIS) in the given EliminateGraph and the number of branches by mis1.

Arguments

eg::EliminateGraph: The graph to compute the MIS for.

Returns

CountingMIS: The size of the maximum independent set.

TensorBranching.counting_mis2 — Method

counting_mis2(eg::EliminateGraph)

Compute the size of the maximum independent set (MIS) in the given EliminateGraph and the number of branches by mis2.

Arguments

eg::EliminateGraph: The graph to compute the MIS for.

Returns

CountingMIS: The size of the maximum independent set.

TensorBranching.cover — Method

cover(sub_covers::AbstractVector{SubCover{INT}}; max_itr::Int = 2, min_complexity::TF = 1.0) where{N, T, TF}

The cover function performs a set covering algorithm on a collection of sub-covers.

Arguments

sub_covers::AbstractVector{SubCover{INT}}: A collection of sub-covers.

max_itr::Int = 2: The maximum number of iterations to perform.

min_complexity::TF = 1.0: The minimum complexity threshold.

Returns

picked: The selected sub-covers.

cx: The complexity of the selected sub-covers.

TensorBranching.covered_by — Method

covered_by(a::LongLongUInt, b::LongLongUInt, mask::LongLongUInt)

Check if a is covered by b with mask. The function returns true if and only if a and b are the same when masked by mask.

TensorBranching.covered_by — Method

covered_by(a::LongLongUInt, clause::Clause)

Check if a is covered by the clause. The function returns true if and only if a and clause.val are the same when masked by clause.mask.

TensorBranching.covered_items — Method

covered_items(bitstrings, clause::Clause)

Return the indices of the bit strings that are covered by the clause.

TensorBranching.effective_γ — Method

effective_γ(branches::Branches, g::SimpleGraph, measure::AbstractMeasure)

Compute the effective γ value for a set of branches in a graph under a given measure.

This function calculates the effective γ value for a given set of branches in a graph. The effective γ value is a measure of complexity that takes into account the size reduction achieved by removing vertices from the graph.

Arguments

branches::Branches: The set of branches for which to calculate the effective γ value.

g::SimpleGraph: The graph for which to calculate the effective γ value.

measure::AbstractMeasure: The measure used to calculate the size reduction.

Returns

The effective γ value for the set of branches.

TensorBranching.graph_from_artifact — Method

graph_from_artifact(num::Int)

Load a graph of the PACE2019 dataset as SimpleGraph.

Arguments

num::Int: The number of the artifact.

Examples

julia> graph_from_artifact(1)
{6160, 40207} undirected simple Int64 graph

TensorBranching.mis_solver — Method

mis_solver(g::SimpleGraph, config::SolverConfig)

Solves the maximum independent set (MIS) problem for a given graph g using a recursive algorithm.

Arguments

g::SimpleGraph: The input graph.

config::SolverConfig: Configuration options for the solver, see SolverConfig.

Returns

max_mis: The size of the maximum independent set.

TensorBranching.neighbor_cover — Method

neighbor_cover(g::SimpleGraph, v::Int, k::Int)

Compute the neighbor cover of a vertex in a graph.

Arguments

g::SimpleGraph: The input graph.

v::Int: The vertex for which to compute the neighbor cover.

k::Int: The number of iterations to perform.

Returns

vertices: An array containing the vertices in the neighbor cover.

openvertices: An array containing the open vertices in the neighbor cover.

TensorBranching.neighbors_2nd — Method

neighbors_2nd(g::SimpleGraph, v::Int)

Return the second-order neighbors of a vertex v in a simple graph g.

Arguments

g::SimpleGraph: The simple graph.

v::Int: The vertex.

Returns

Array{Int}: An array of second-order neighbors of v.

TensorBranching.open_neighbors — Method

open_neighbors(g::SimpleGraph, vertices::Vector{Int})

Returns a vector of vertices in the graph g, which are neighbors of the given vertices and not in the given vertices.

Arguments

g::SimpleGraph: The graph in which to find the open neighbors.

vertices::Vector{Int}: The vertices for which to find the open neighbors.

Returns

A vector of open neighbors of the given vertices.

TensorBranching.open_vertices — Method

open_vertices(g::SimpleGraph, vertices::Vector{Int})

Remove vertices from the given vector that are connected to all other vertices in the graph.

Arguments

g::SimpleGraph: The graph object.

vertices::Vector{Int}: The vector of vertices.

Returns

Vector{Int}: The open vertices.

TensorBranching.optimal_branches — Method

optimal_branches(g::SimpleGraph, vertices::AbstractVector, strategy::AbstractBranching; measure::AbstractMeasure = NumOfVertices(), filter::AbstractTruthFilter = NoFilter())

Find the optimal branches for the given graph and vertices.

Arguments

g::SimpleGraph: the input graph

vertices::AbstractVector: the vertices to be considered

strategy::AbstractBranching: the branching strategy, e.g., NaiveBranching(), SetCoverBranching()

Keyword Arguments

measure::AbstractMeasure: the measure for the branching strategy, e.g., NumOfVertices(), D3Measure()

filter::AbstractTruthFilter: the filter for the branching strategy, e.g., EnvFilter(), NoFilter()

table_solver: the solver configuration, e.g., TensorNetworkSolver()

TensorBranching.removed_vertices — Method

removed_vertices(vertices::Vector{Int}, g::SimpleGraph, clause::Clause{N}) where N

Given a list of vertices, a graph, and a clause, this function returns a list of removed vertices.

The vertices argument is a vector of integers representing the vertices to consider. The g argument is a SimpleGraph object representing the graph. The clause argument is a Clause object representing a clause.

The function iterates over the vertices and checks if the corresponding bit in the clause.mask is 1. If it is, the vertex is added to the list of removed vertices (rvs). If the corresponding bit in the clause.val is also 1, the neighbors of the vertex are also added to rvs.

The function returns the list of removed vertices with duplicates removed.

TensorBranching.setcover_strategy — Method

setcover_strategy(tbl::BranchingTable{INT}, vertices::Vector{Int}, g::SimpleGraph, max_itr::Int, measure::AbstractMeasure, setcover_solver::AbstractSetCoverSolver) where{INT}

This function implements a set covering strategy for the optimal branching in the following steps.

Compute the sub-covers of the branching table.

A set of sub-covers with minimum effective γ is selected by the set covering algorithm.

The selected sub-covers are converted to branching rules.

Arguments

tbl::BranchingTable{INT}: The branching table.

vertices::Vector{Int}: The vertices to be covered.

g::SimpleGraph: The graph.

max_itr::Int: The maximum number of iterations.

measure::AbstractMeasure: The measure used to evaluate the quality of a cover.

setcover_solver::AbstractSetCoverSolver: The set cover solver, linear programming or integer programming.

Returns

branches: The branches generated by the set covering strategy.

TensorBranching.solve_mis — Method

solve_mis(g::SimpleGraph, config::SolverConfig)

Solve the maximum independent set problem for the given graph and configuration.

Arguments

g::SimpleGraph: the input graph

config::SolverConfig: the solver configuration

TensorBranching.subcovers — Method

subcovers(bss::AbstractVector{Vector{INT}}, vertices::Vector{Int}, g::SimpleGraph, use_rv::Bool) where {INT}

Compute the subcovers of a set of bit strings by iteratively gathering clauses.

Arguments

bss::AbstractVector{Vector{INT}}: A vector of vectors of bit strings.

vertices::Vector{Int}: A vector of integers representing vertices.

g::SimpleGraph: A simple graph.

use_rv::Bool: A boolean indicating whether to use the number of vertices removed.

Returns

allcovers::Vector{SubCover}: A vector of SubCover objects representing the subcovers.

TensorBranching.subcovers_naive — Method

subcovers_naive(bs, vertices, g, use_rv)

Compute the subcovers for a given set of bit strings by search all possible clauses.

Arguments

bs::Union{Vector{INT}, AbstractVector{Vector{INT}}}: The set of bit strings.

vertices::Vector{Int}: The vertices.

g::SimpleGraph: The graph.

use_rv::Bool: A flag indicating whether to use the number of vertices removed.

Returns

allcovers::Vector{SubCover{INT}}: The computed subcovers.