API

Index

TensorBranching.BranchingNode

TensorBranching.BranchingTable

TensorBranching.Clause

TensorBranching.D3Measure

TensorBranching.NaiveBranching

TensorBranching.SetCoverBranching

TensorBranching.SolverConfig

TensorBranching.SubCover

TensorBranching.branching_tree

TensorBranching.clause

TensorBranching.complexity

TensorBranching.counting_mis1

TensorBranching.counting_mis2

TensorBranching.cover

TensorBranching.covered_by

TensorBranching.covered_by

TensorBranching.covered_items

TensorBranching.missolve

TensorBranching.optimal_branches

TensorBranching.subcovers

TensorBranching.subcovers_naive

Modules

TensorBranching.BranchingNode — Type

mutable struct BranchingNode

A mutable struct representing a node in a branching tree.

Fields

children::Vector{BranchingNode}: The children nodes of the current node.

graph::SimpleGraph{Int}: The graph associated with the current node.

removed::Vector{Vector{Int}}: The removed edges from the graph.

Constructor

BranchingNode(graph::SimpleGraph{Int}; children::Vector{BranchingNode} = Vector{BranchingNode}(), removed::Vector{Vector{Int}} = Vector{Vector{Int}}()): Constructs a new BranchingNode object.

TensorBranching.BranchingTable — Type

BranchingTable{N, C}

A table of branching configurations. The table is a vector of vectors of StaticBitVector{N, C}. Type parameters are:

N: The number of bits in the bit vector.

C: The number of integers as the storage.

Examples

julia> graph_sat = graph_from_tuples(3, [(1, 2), (2, 3), (1, 3)])
{3, 3} undirected simple Int64 graph

julia> tbl = reduced_alpha_configs(TensorNetworkSolver(), graph_sat, [1, 2])
BranchingTable{N}
001

To cover the branching table, at least one clause in each row must be satisfied.

TensorBranching.Clause — Type

Clause{N, T}

A clause is a pair of bit strings, mask and val, where mask is a bit string that indicates the bits that are relevant to the clause, and val is a bit string that indicates the bits that must be satisfied. The clause is satisfied if and only if val is covered by the bit string and mask.

N: The number of bits in the bit vector.

T: The number of integers as the storage.

Examples

julia> Clause(bit"1110", bit"1001")
Clause{4, 3, Int64}: mask: 1110 ₍₂₎, val: 1000 ₍₂₎

If some bit in mask is set as 0, then the corresponding bit in val must be 0.

TensorBranching.D3Measure — Type

D3Measure <: AbstractMeasure

A measure of complexity by counting the number of vertices with degree at least 3.

TensorBranching.NaiveBranching — Type

struct NaiveBranching <: AbstractBranching

A struct representing the NaiveBranching branching strategy.

TensorBranching.SetCoverBranching — Type

struct SetCoverBranching <: AbstractBranching

A struct representing a branching strategy for set cover problems.

Fields

max_itr::Int: The maximum number of iterations.

Constructors

SetCoverBranching(): Constructs a SetCoverBranching object with a default value of 2 for max_itr.

SetCoverBranching(max_itr::Int): Constructs a SetCoverBranching object with the specified max_itr value.

TensorBranching.SolverConfig — Type

SolverConfig{ST<:AbstractMISSolver, BT<:AbstractBranching, MT<:AbstractMeasure, VT<:AbstractVertexSelector, FT<:AbstractTruthFilter}

The configuration for the optimal branching MIS solver.

Fields

table_solver::ST: the MIS solver, e.g., TensorNetworkSolver()

branching_strategy::BT: the branching strategy, e.g., NaiveBranching(), SetCoverBranching()

measurement::MT: the measurement for the branching strategy, e.g., NumOfVertices(), D3Measure()

vertex_selector::VT: the vertex selector, e.g., MinBoundarySelector(2), ManualSelector([1, 2, 3])

table_filter::FT: the truth filter, e.g., EnvFilter(), NoFilter()

TensorBranching.SubCover — Type

SubCover{N, T}

A subcover is a pair of a set of integers ids, a clause clause and a integer n_rm. The ids for the truth covered by the clause, and n_rm is the number of vertices to remove.

N: The number of bits in the bit vector.

T: The number of integers as the storage.

Examples

julia> SubCover([1, 2], Clause(bit"1110", bit"1001"), 3)
SubCover{4, Int64}: ids: Set([2, 1]), mask: 1110 ₍₂₎, val: 1000 ₍₂₎, n_rm: 3

TensorBranching.branching_tree — Method

branching_tree(g::SimpleGraph, strategy::AbstractBranching, kneighbor::Int, use_rv::Bool)

Constructs a branching tree based on the given graph g using the specified strategy. The kneighbor parameter determines the number of neighbors to consider during branching. If use_rv is true, number of vertex removed is used; otherwise, only count the clause.

Arguments

g::SimpleGraph: The input graph.

strategy::AbstractBranching: The branching strategy to use.

kneighbor::Int: The number of neighbors to consider during branching.

use_rv::Bool: Whether to use number of vertex removed.

Returns

tree: The constructed branching tree.

branch_num: The number of branches in the tree.

TensorBranching.clause — Method

clause(bitstrings::AbstractVector{BitStr{N, T}})

Return a clause that covers all the bit strings.

TensorBranching.complexity — Method

complexity(sbranches::Vector{Int})

Compute the complexity of a set of branches.

This function solves the equation 1 = sum(x^(-i) for i in sbranches) for x, where sbranches is a vector of integers representing the branches.

Arguments

sbranches::Vector{Int}: A vector of integers representing the branches.

Returns

Float64: The value of x that satisfies the equation.

TensorBranching.counting_mis1 — Method

counting_mis1(eg::EliminateGraph)

Compute the size of the maximum independent set (MIS) in the given EliminateGraph and the number of branches by mis1.

Arguments

eg::EliminateGraph: The graph to compute the MIS for.

Returns

CountingMIS: The size of the maximum independent set.

TensorBranching.counting_mis2 — Method

counting_mis2(eg::EliminateGraph)

Compute the size of the maximum independent set (MIS) in the given EliminateGraph and the number of branches by mis2.

Arguments

eg::EliminateGraph: The graph to compute the MIS for.

Returns

CountingMIS: The size of the maximum independent set.

TensorBranching.cover — Method

cover(sub_covers::AbstractVector{SubCover{N, T}}; max_itr::Int = 2, min_complexity::TF = 1.0) where{N, T, TF}

The cover function performs a set covering algorithm on a collection of sub-covers.

Arguments

sub_covers::AbstractVector{SubCover{N, T}}: A collection of sub-covers.

max_itr::Int = 2: The maximum number of iterations to perform.

min_complexity::TF = 1.0: The minimum complexity threshold.

Returns

picked: The selected sub-covers.

cx: The complexity of the selected sub-covers.

TensorBranching.covered_by — Method

covered_by(a::BitStr, b::BitStr, mask::BitStr)

Check if a is covered by b with mask. The function returns true if and only if a and b are the same when masked by mask.

TensorBranching.covered_by — Method

covered_by(a::BitStr, clause::Clause)

Check if a is covered by the clause. The function returns true if and only if a and clause.val are the same when masked by clause.mask.

TensorBranching.covered_items — Method

covered_items(bitstrings, clause::Clause)

Return the indices of the bit strings that are covered by the clause.

TensorBranching.missolve — Method

missolve(g::SimpleGraph, config::SolverConfig; show_count = false)

Solve the maximum independent set problem for the given graph and configuration.

Arguments

g::SimpleGraph: the input graph

config::SolverConfig: the solver configuration

Keyword Arguments

show_count::Bool: whether to show the count of the MIS

TensorBranching.optimal_branches — Method

optimal_branches(g::SimpleGraph, vertices::AbstractVector, strategy::AbstractBranching; measurement::AbstractMeasure = NumOfVertices(), filter::AbstractTruthFilter = NoFilter())

Find the optimal branches for the given graph and vertices.

Arguments

g::SimpleGraph: the input graph

vertices::AbstractVector: the vertices to be considered

strategy::AbstractBranching: the branching strategy, e.g., NaiveBranching(), SetCoverBranching()

Keyword Arguments

measurement::AbstractMeasure: the measurement for the branching strategy, e.g., NumOfVertices(), D3Measure()

filter::AbstractTruthFilter: the filter for the branching strategy, e.g., EnvFilter(), NoFilter()

table_solver: the solver configuration, e.g., TensorNetworkSolver()

TensorBranching.subcovers — Method

subcovers(bss::AbstractVector{Vector{BitStr{N, T}}}, vertices::Vector{Int}, g::SimpleGraph, use_rv::Bool) where {N, T}

Compute the subcovers of a set of bit strings by iteratively gathering clauses.

Arguments

bss::AbstractVector{Vector{BitStr{N, T}}}: A vector of vectors of bit strings.

vertices::Vector{Int}: A vector of integers representing vertices.

g::SimpleGraph: A simple graph.

use_rv::Bool: A boolean indicating whether to use the number of vertices removed.

Returns

allcovers::Vector{SubCover}: A vector of SubCover objects representing the subcovers.

TensorBranching.subcovers_naive — Method

subcovers_naive(bs, vertices, g, use_rv)

Compute the subcovers for a given set of bit strings by search all possible clauses.

Arguments

bs::Union{Vector{BitStr{N, T}}, AbstractVector{Vector{BitStr{N, T}}}}: The set of bit strings.

vertices::Vector{Int}: The vertices.

g::SimpleGraph: The graph.

use_rv::Bool: A flag indicating whether to use the number of vertices removed.

Returns

allcovers::Vector{SubCover{N, T}}: The computed subcovers.